+++ data/samson.txt
@@ -1,8 +1,6 @@
 T | 0 | read book
-D | 0 | return book | 2019-10-15 1600
 D | 0 | read book | 2024-12-30 1300
 T | 0 | read book
 D | 0 | drink water | 2024-07-29 1234
 T | 0 | read book
 T | 0 | write_task
-T | 0 | read_book2
