+++ src/main/java/samson/Greeting.java
@@ -1,3 +1,5 @@
+package samson;
+
 public class Greeting {
     private String chatbotName;
