+++ src/main/java/samson/command/AddCommand.java
@@ -1,5 +1,9 @@
-// AddCommand.java
+package samson.command;// Samson.Samson.Samson.Samson.Command.Command.AddCommand.java
 import java.io.IOException;
+
+import samson.Storage;
+import samson.Ui;
+import samson.task.*;
 
 public class AddCommand extends Command {
     private Task task;
