+++ src/main/java/samson/command/Command.java
@@ -1,3 +1,9 @@
+package samson.command;
+
+import samson.Storage;
+import samson.Ui;
+import samson.task.*;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws Exception;
     public abstract boolean isExit();
