+++ src/main/java/samson/task/Deadline.java
@@ -1,6 +1,10 @@
+package samson.task;
+
 import java.time.DateTimeException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+
+import samson.SamException;
 
 public class Deadline extends Task {
     private LocalDateTime by;
