+++ src/main/java/samson/task/Event.java
@@ -1,6 +1,10 @@
+package samson.task;
+
 import java.time.DateTimeException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+
+import samson.SamException;
 
 public class Event extends Task {
     private LocalDateTime from;
