+++ src/main/java/samson/task/Task.java
@@ -1,3 +1,4 @@
+package samson.task;
 
 public abstract class Task {
     private String description;
