+++ src/main/java/samson/task/TaskType.java
@@ -1,3 +1,5 @@
+package samson.task;
+
 public enum TaskType {
     TODO,
     DEADLINE,
