+++ src/main/java/samson/task/ToDo.java
@@ -1,4 +1,6 @@
-public class ToDo extends Task{
+package samson.task;
+
+public class ToDo extends Task {
     public ToDo(String description) {
         super(description, TaskType.TODO);
     }
